\name{TDA-internal}

\alias{findKtree}
\alias{gridBy}
\alias{plotRule}
\alias{rMultinom}
\alias{siloF}
\alias{AlphaComplexDiagGUDHI}
\alias{AlphaShapeDiagGUDHI}
\alias{Bottleneck}
\alias{Dtm}
\alias{DtmWeight}
\alias{GridDiag}
\alias{Kde}
\alias{KdeDist}
\alias{RipsDiag}
\alias{Wasserstein}

\title{
  Internal TDA functions
}

\description{
  Internal TDA functions
}

\details{
  These are functions not to be called by the user, including functions generated by Rcpp.
}

\keyword{internal}