% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDA_Kendall.R
\name{TDA_Kendall}
\alias{TDA_Kendall}
\title{Complete TDA power analyis}
\usage{
TDA_Kendall(conc, time)
}
\arguments{
\item{conc}{vector of conc concentration, BIBI, etc.}

\item{time}{vector of time}

\item{site}{vector of sites}
}
\value{
A list of trends.
}
\description{
Complete TDA power analyis
}
\examples{
TDA_Kendall(1, 1)
}
