% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDA.R
\docType{package}
\name{TDA-package}
\alias{TDA}
\alias{TDA-package}
\title{TDA: (Trend Detection Analysis) some functions for implement McBride 2019
\description{
TDA provides two function for implementing the trend detection analysis tool described in McBride 2019
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/clarkbar88/TDA}
}

}
