\name{TDA-package}
\alias{TDA-package}
\alias{TDA}
\docType{package}
\title{
Statistical Tools for Topological Data Analysis
}
\description{
This package provides some tools for Topological Data Analysis. In particular it provides functions for the statistical analysis of persistent homology and for density clustering.
}
\details{
\tabular{ll}{
Package: \tab TDA\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2014-08-29\cr
License: \tab GPL-3 \cr
}
}
\author{
Brittany Terese Fasy, Jisu Kim, Fabrizio Lecci, Clement Maria

Maintainer: Fabrizio Lecci <lecci@cmu.edu>
}
\references{
Herbert Edelsbrunner, and John Harer, (2010), "Computational topology: an introduction". American Mathematical Society.

Brittany T. Fasy, Fabrizio Lecci, Alessandro Rinaldo, Larry Wasserman, Sivaraman Balakrishnan, and Aarti Singh. (2013), "Statistical Inference For Persistent Homology: Confidence Sets for Persistence Diagrams", (arXiv:1303.7117). To appear, Annals of Statistics.

Chazal F, Fasy BT, Lecci F, Michel B, Rinaldo A, Wasserman L (2014). "Robust Topological Inference: Distance-To-a-Measure and Kernel Distance." Technical Report.	

Frederic Chazal, Brittany T. Fasy, Fabrizio Lecci, Alessandro Rinaldo, and Larry Wasserman, (2014), "Stochastic Convergence of Persistence Landscapes and Silhouettes", Proceedings of the 30th Symposium of Computational Geometry (SoCG). (arXiv:1312.0308)

Frederic Chazal, Brittany T. Fasy, Fabrizio Lecci, Bertrand Michel, Alessandro Rinaldo, and Larry Wasserman, (2014), "Subsampling Methods for Persistent Homology". (arXiv:1406.1901)

Clement Maria, "GUDHI, Simplicial Complexes and Persistent Homology Packages". https://project.inria.fr/gudhi/software/.

Dmitriy Morozov, "Dionysus, a C++ library for computing persistent homology". http://www.mrzv.org/software/dionysus/

Ulrich Bauer, Michael Kerber, Jan Reininghaus, "PHAT, a software library for persistent homology". https://code.google.com/p/phat/
}
\keyword{ package }
