% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDA_Kendall.R
\name{TDA_Kendall}
\alias{TDA_Kendall}
\title{Complete TDA power analyis}
\usage{
TDA_Kendall(conc, time)
}
\arguments{
\item{conc}{vector of conc (concentration, BIBI, etc.)}

\item{time}{vector of time}

\item{site}{vector of sites}
}
\value{
A tibble of trends and power.
}
\description{
Complete TDA power analyis
}
\examples{
data(all_bibi_data)
tmp<-subset(all_bibi_data,Site=='08BEA3312')
TDA_Kendall(tmp$`Overall Score`,tmp$Year)
}
